+++ src/Exception/Config/ConfigAlreadyExistsException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PhpZone\PhpZone\Exception\Config;
+
+class ConfigAlreadyExistsException extends \Exception
+{
+}